backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.flake8]
max-line-length = 88

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.110.1"
pydantic-extra-types = "^2.6.0"
pymongo = "^4.6.3"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
